Add a layout for the landing page
