--- web-app/src/app/(landing)/components/auth-buttons.tsx
@@ -1,16 +1,0 @@
-import Link from "next/link";
-
-import { Button } from "@/components/ui/button";
-
-export function AuthButtons() {
-  return (
-    <div className="flex items-center gap-4">
-      <Link href="/signin">
-        <Button variant="outline">Sign In</Button>
-      </Link>
-      <Link href="/signup">
-        <Button>Sign Up</Button>
-      </Link>
-    </div>
-  );
-}
