+++ web-app/src/app/(landing)/components/footer.tsx
@@ -1,9 +1,13 @@
 import Link from "next/link";
 
 import { SokosumiLogo } from "@/components/masumi-logos";
+
 export default function Footer() {
   return (
-    <footer className="bg-landing-footer-background text-landing-footer-foreground">
+    <footer
+      id="footer"
+      className="bg-landing-footer-background text-landing-footer-foreground"
+    >
       <div className="container mx-auto px-4 py-12">
         {/* Top section with logo */}
         <div className="mb-8">
