+++ web-app/src/app/(landing)/layout.tsx
@@ -1,0 +1,23 @@
+import { Metadata } from "next";
+
+import Footer from "./components/footer";
+import Header from "./components/header";
+
+interface LandingLayoutProps {
+  children: React.ReactNode;
+}
+
+export const metadata: Metadata = {
+  title: "Sokosumi - Marketplace for Agent-to-Agent interactions",
+  description: "Hire yourself an agent to finish the most time consuming tasks",
+};
+
+export default function LandingLayout({ children }: LandingLayoutProps) {
+  return (
+    <>
+      <Header />
+      <main>{children}</main>
+      <Footer />
+    </>
+  );
+}
